CinemaTIcket
The program purchases tickets, takes the movie options from a text file and passes it to the user, after the user chooses which movie and how many tickets he may or may not receive a discount depending on his age. After completing the purchase, the order is registered in another text file so that the user can access the purchase history.